+++ src/roslibpy/__init__.py
@@ -90,24 +90,50 @@ two integers: seconds since epoch and nanoseconds since seconds.
 
 """
 
-from .__version__ import __author__
-from .__version__ import __author_email__
-from .__version__ import __copyright__
-from .__version__ import __description__
-from .__version__ import __license__
-from .__version__ import __title__
-from .__version__ import __url__
-from .__version__ import __version__
-from .core import Header
-from .core import Message
-from .core import Param
-from .core import Service
-from .core import ServiceRequest
-from .core import ServiceResponse
-from .core import Time
-from .core import Topic
-from .ros import Ros
-from .ros import set_rosapi_timeout
+from .__version__ import (
+    __author__,
+    __author_email__,
+    __copyright__,
+    __description__,
+    __license__,
+    __title__,
+    __url__,
+    __version__
+)
+from .core import (
+    Header,
+    Message,
+    Param,
+    Service,
+    ServiceRequest,
+    ServiceResponse,
+    Time,
+    Topic
+)
+from .ros import (
+    set_rosapi_timeout,
+    Ros
+)
 
-__all__ = ['Ros', 'set_rosapi_timeout', 'Message', 'Header', 'Time', 'Param', 'Service', 'ServiceRequest', 'ServiceResponse', 'Topic',
-           '__author__', '__author_email__', '__copyright__', '__description__', '__license__', '__title__', '__url__', '__version__']
+__all__ = [
+    '__author__',
+    '__author_email__',
+    '__copyright__',
+    '__description__',
+    '__license__',
+    '__title__',
+    '__url__',
+    '__version__',
+
+    'Header',
+    'Message',
+    'Param',
+    'Service',
+    'ServiceRequest',
+    'ServiceResponse',
+    'Time',
+    'Topic',
+
+    'set_rosapi_timeout',
+    'Ros',
+]