+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Unreleased
 
 * Fixed #87 where a goal could be marked as terminal on result alone rather
   than both result and status.
+* Ensure input of ``Time`` is always two integers.
 
 **Deprecated**